week11/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dnspython==2.4.2
fastapi==0.104.1
fonttools==4.45.0
h11==0.14.0
httptools==0.6.1
idna==3.4
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
paho-mqtt==1.6.1
pandas==2.1.3
Pillow==10.1.0
pydantic==2.5.1
pydantic_core==2.14.3
pymongo==4.6.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
setuptools==68.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
wheel==0.41.2
